Eliminate setjmp() from getopt
Instead of using longjmp to return to the start of the switch,
use continue to restart the loop, going through getopt() several
times. It repeatedly returns "(dummy)" while initialization is
in progress.

Avoiding setjmp() avoids a lot of signal mask save/restore syscalls.

This patch also makes the initialization stages more explicit. There
are a couple of temporary bodges to minimize the size of the patch
and to make it clear the existing tests still pass.