Add ttyfd: Returns a descriptor to the terminal
Use ctermid+open to get a descriptor to the connected terminal; or
if that fails, duplicate the first of (stderr, stdin, stdout) which
is a terminal.

Note that this will fail if we cannot open the terminal (e.g., we
are running inside a chroot without /dev/) and all of the standard
file descriptors have been redirected.