Rewrite SHA256 round computation
The old round computation used two temporary variables in order to
convey to the compiler a suggestion as to the most efficient order
in which to perform the necessary operations.  A decade after I
wrote that code, advances in compilers and CPUs have made those
hints unnecessary; and having two temporary values explicitly held
on the stack is undesireable for security reasons.

On {clang 3.8.0, gcc 4.8.5}/{-O2, -O3 -march=native}, the performance
impact of this is not measurably different from zero.