Add workaround for bash environment mangling
The shell builtin "command -p foo" executes the command "foo" after finding
it within a set of default system paths.  This is equivalent to setting
$PATH to a standard value and then running "foo", except that the program
should be executed with $PATH unchanged.  In other words,
	export PATH=foobar
	command -p sh -c 'echo $PATH'
should print "foobar" (the value of $PATH when "command" was invoked), not
the default system path used to find "sh".

The bash shell does this wrong, resulting in us losing the $PATH provided by
the system make logic.  In particular if the compiler binary is found under
/usr/local/bin but the default system path is /usr/bin:/bin:/usr/sbin:/sbin,
our portability-workaround scripts would find that the compiler could not be
executed.

To work around this bug, we provide the (original) $PATH as a command-line
argument; and if it does not match $PATH we print a warning and reset $PATH
to the value passed on the command-line.

This change requires users of libcperciva/POSIX/posix-*.sh to be modified in
order to add the new "$PATH" argument.