Explicit casts to (size_t)
For the below, define:
    struct pollfds * current = &fds[fdscanpos];

These casts are safe assuming that current->fd is only set to -1
if (current->revents & POLLNVAL); in other words, no other piece
of code will change the value of ->fd).  This is true at the
moment, and we can probably (?) rely on nobody changing the code
to break this in the future.

I added the "continue" to make sure that even if the POLLNVAL and
(POLLIN | POLLOUT) flags were not mutually exclusive[1], we
wouldn't have any problems.

[1] the Linux man page states:
	The bits returned in revents can include any of those
	specified in events, or one of the values POLLERR,
	POLLHUP, or POLLNVAL.
        http://man7.org/linux/man-pages/man2/poll.2.html
If we interpret the first "or" as "xor", then this would give us
the needed assurance.

The POSIX standard doesn't include language exactly like that, but
it does say
    POLLIN         Data other than high priority data may be read
                   without blocking.
    ...
    POLLNVAL       Invalid fd member (revents only).
which I suppose implicitly means that if POLLNVAL is set, POLLIN
cannot be set.