Switch events_network from select to poll
This has several effects: Most notably, it removes the FD_SETSIZE
limitation on descriptors, but it will also improve performance in
the unlikely case of waiting on a very sparse set of descriptors
(since poll is only passed the active list), and it may yield a
modest speedup in general (since the list is no longer regenerated
on each call to events_network_select, although the kernel still
needs to examine the entire list).

Note that this change is likely to result in events being returned
in a different order if multiple descriptors become ready at the
same time; this may expose previously-hidden bugs.

This change adds a dependency on ctassert.h; projects which do not
already include that file will need to add it.