parsenum: disable clang -Wconversion and friends
Depending on the version of clang, we either need to disable warnings
on conversions generally or on float, sign, and 64-to-32 conversions
separately.  Because we do all of these, we also need to disable
warnings about disabling warnings which don't exist.

We can't disable warnings on gcc, due to a bug which breaks _Pragma
inside macros.