Add PARSENUM(x, s)
For floating-point and unsigned integer types, PARSENUM(x, s) functions
as PARSENUM(x, s, min, max) where min and max are set based on the
limits of the variable.

Calling the two-argument form of PARSENUM with a signed integer variable
is an assertion failure, since it is impossible to safely determine the
limits of a signed integer type.