Rework range checks in unsigned case
There's no need to include range checks in the PARSENUM macro which
are already handled within the parsenum_unsigned function; all we
need to do is check that max >= 0 (since it would get mangled by
casting to unsigned) and adjust the range-check to ask for values
between MAX(min, 0) and MIN(max, (*(x) = -1)).

While we're here, adjust the check for max < 0 so that running
PARSENUM(size_t, "foobar", -2, -1) will set EINVAL due to "foobar"
not being parsable rather than setting ERANGE due to [-2, -1] not
being satisfiable by an unsigned value.