Avoid multiple-definition issues in CTASSERT
Redefining a typedef, even if the definition hasn't changed, is not
allowed in C99.  As a result, CTASSERT could cause problems if used
in header files, since the __LINE__ numbers could collide.  (Also,
but far less likely to occur by accident: If someone wrote
    CTASSERT(1); CTASSERT(2);
or equivalent code.)

Switch from defining a *typedef* for each compile-time assertion to
declaring an *external array*.  Since extern symbols can be declared
multiple times, this avoids the collision problem; but the same rules
about arrays not having negative size ensures that the assertion is
checked.