test_getopt: move test logic from Makefile to .sh
This is commit 2/2 for preparing the getopt test for a new build system.
These commits logically belong together, but I separated them so that it
minimizes the diffs and makes it easier to see what's happening.

The end goal (which prompted this change) is to:
1) have the Makefile run a single shell script for `make test`
2) preserve the existing behaviour

However, if we put the entire test logic in a single shell script, we'd
get different behaviour.  Namely,
  /bin/sh -e test_getopt_args.sh 2>test_getopt.log || true
will stop running test_getopt_args.sh as soon as one of the commands
within it has a non-zero exit code, whereas
  run_tests 2>test_getopt.log || true
(where "run_tests" is a function) will continue running commands in the
function after a non-zero exit code, even when the script is invoked
with -e.

Solution:
- rename the old test_getopt.sh to something else
- move the Makefile logic into a new test_getopt.sh, which calls the
  old test_getopt.sh in exactly the same manner that the Makefile did.

(the reason to I'm renaming the old test_getopt.sh is so that we can
standardize on test_*.sh being the main test file)