build: add tests/{buildall,monoclock,sha256}/Makefile.BSD
I'm doing this separately because these each have a Makefile.inc1 file.

Note that as a result of this change, we can no longer simply cd into the
relevant test/*/ directory and run `make` -- we won't have the required
${CFLAGS_POSIX} and ${LDADD_POSIX}.  Instead, we rely on the user running
`make` from the top-level directory.  Since this is how all the other Tarsnap
software works, I think this is an acceptable change.