tests: use valgrind for complicated tests
By sourcing (`.`) the subdirectory scripts, we pass in the ${c_valgrind_cmd}
variable, which will either be a real command or empty (depending on whether
the tests are using valgrind).  If the script is called directly, then
${c_valgrind_cmd} will be undefined and treated as empty[1].

[1] Provided that the user did not deliberately add
    set -o nounset
to her login shell, which would likely ruin a whole bunch of other scripts.

The main drawback of this method is that we can only run the valgrind command
once in the subdirectory script.  If we wanted to do more memory-testing, I
suggest that we add the command to the relevant tests/??-*.sh file (after
running setup_check_variables, of course).