tests: don't use set -e
We want the main test framework to print PASS / FAIL, and output the tail of
the valgrind log if relevant.  Using `set -e` on the main script meant that we
could only print PASS, as a failure would immediately exit.

Instead of sourcing a subshell, we need to run
    c_valgrind_cmd=${c_valgrind_cmd} ./test_PROGNAME.sh
and that shell script can `set -e` if desired.