sha256_shani.c: cast consts to (int64_t) for _mm_set_epi64x()
On my system, _mm_set_epi64x() wants two (long long) args.  However, clang
complains if the argument is >= 0x8000000000000000, because it assumes that
it's unsigned, but implicitly casting it to signed makes it negative.

Quite *why* _mm_set_epi64x() thinks it wants signed values is beyond my
understanding.  Presumably this is Intel's fault in the SSE2 API?