tests: keep valgrind failure exitcode for setuidgid
Not a huge deal -- a failure would still give us an exit code of 1.  But having
the real value would trigger the test suite to display memory info, rather than
a generic "FAILED!".

Why does this matter?  Well, Ubuntu 18.04 has a memory leak in nss_load_library
when we have
    ./tests/setuidgid/test_setuidgid 1001
but not
    ./tests/setuidgid/test_setuidgid td

That's going to be just _fascinating_ to track down, I can tell.