cpusupport: fix for detection of SHA extensions
It looks like __get_cpuid() can only used for the basic CPUID feature bits.  To
detect the extended features (EAX=7, ECX=0), we want to end up with:

    asm volatile ("cpuid"
         :"=a"(a), "=b"(b), "=c"(c), "=d"(d)
         :"a"(a), "c"(c)
        );

but __get_cpuid() calls __cupid(), which is:

    #define __cpuid(level, a, b, c, d)                      \
      __asm__ ("cpuid\n\t"                                  \
               : "=a" (a), "=b" (b), "=c" (c), "=d" (d)     \
               : "0" (level))

Note the lack of the final
    "c"(c)

And yes, missing that (c <- 0) makes ebx contain 0, instead of a sensible value
like 571761281 (with the Intel emulator doing Goldmont).  And no, setting
    eax=7; ecx=0;
before calling __get_cpuid() didn't help.