tests/parsnum: deal with drunken clang & glibc on Ubuntu 18.04
Without this, we simultaneously have "isnan() is converting to float"
and "isnan() is converting to long double":

    main.c:17:12: error: implicit conversion loses floating-point precision: 'double' to 'float' [-Werror,-Wconversion]
            if (isnan(x) && isnan(y))
                ~~~~~~^~
    /usr/include/math.h:644:46: note: expanded from macro 'isnan'
    #  define isnan(x) __MATH_TG ((x), __isnan, (x))
                       ~~~~~~~~~~~~~~~~~~~~~~~~~~^~~
    /usr/include/math.h:559:16: note: expanded from macro '__MATH_TG'
       ? FUNC ## f ARGS                             \
         ~~~~~~~~~ ^~~~
    main.c:17:12: error: implicit conversion increases floating-point precision: 'double' to 'long double' [-Werror,-Wdouble-promotion]
            if (isnan(x) && isnan(y))
                ~~~~~~^~
    /usr/include/math.h:644:46: note: expanded from macro 'isnan'
    #  define isnan(x) __MATH_TG ((x), __isnan, (x))
                       ~~~~~~~~~~~~~~~~~~~~~~~~~~^~~
    /usr/include/math.h:562:16: note: expanded from macro '__MATH_TG'
       : FUNC ## l ARGS)
         ~~~~~~~~~ ^~~~

On Ubuntu 18.04, /usr/include/math.h contains 54 lines of definitions
for __MATH_TG (with lots of #if for various lengths of floats).  From
reading various theories on mailing lists and github issues, it looks
like clang is evaluating all the branches of isnan() and related macros.
Including branches not taken, such as "if the data type is a double, use
the double version of isnan() instead of the float version.".  :(

(The same applies to isinf(), signbit(), and fpclassify() as well.)