apisupport: add linux getrandom()
Although getrandom() was added in Linux 3.17 (2014-10), it was only
added to glibc in 2.25 (2017-02), and Linux distros are cautious about
upgrading to a new libc.  In those cases, getrandom() is only available
if you use syscall(2) to do an indirect system call.

For example, Debian 9 "stretch" (2017-06) still uses 2.24, whereas
Ubuntu 18.04 has glibc 2.27.