tests/buildsingles: Makefile.inc1 -> Makefile-headers
This is to ensure that "make publish" distributes this file.  (Sure, we don't
distribute libcperciva as a tarball, but it's good to have the framework sorted
out.)

The "make publish" framework assumes that any Makefile\.* is only used for
updating the real (posix) Makefiles and thus excludes them from the tarball.
And other than tests/buildsingles/Makefile.inc1, that's valid -- we don't want
to distribute Makefile.BSD, Makefile.inc, and Makefile.prog.

However, we can't compile the "buildsingles" tests without this file;
consequently, if we generated a theoretical libcperciva-VERSION.tar.gz, it
would be unable to "make".