monoclock: strict usage of #if define
This should have been part of 518aa7487a09673bdf045ecd583530388c35ab87.

Some versions of OSX don't supply CLOCK_REALTIME (a POSIX violation),
and as a result we fall back to gettimeofday().  In the regular
monoclock_get(), this command can fail, so we need err0.  However,
there's no posibility of error in monoclock_getres() if we don't have
CLOCK_REALTIME.