metabuild: move Makefile generation into a sh script
It's grown beyond what can conveniently be done in a Makefile -- it's
awkward dealing with the special escape characters, changing directory
all the time, and being unable to add comments in the middle of the
processing.

This moves the Makefile generation into a separate script, and makes
minimal changes to the code.  Simplifications will come in the next few
commits.

This commit does not change any functionality.