tests/parsenum: special MacOS X error-handling for abort()
MacOS X has some special handling for abort... despite having the bang in
    ! ./test_parsenum
for the abort-case, the shell script still bails.

Adding the
    set +e; ! ./test_parsnum
abolishes the test that the abort-cases produce a non-zero error code,
but I think that grepping for the specific error message is enough to
verify that PARSENUM is doing the right thing.