tests/mpool: take median time instead of lowest
We originally went with the lowest time because that seemed like a
robust estimate of the time without interruptions -- the task can't
complete faster than it completes, so in theory anything over the lowest
value indicates delay due to system load.

However, some systems have a relatively poor resolution of the
CLOCK_PROCESS_CPUTIME_ID clock; I'm seeing a few instances of 0-duration
tests.  Using the median allows this benchmark to be robust against a
few 0s.  Of course, if more than half the executions result in 0
duration, then we'll need to tweak the number of repetitions and/or try
other solutions.