tests/mpool: stricter bound on mpool performance
On systems with no load, I consistently get mpool being 96% - 98% as
fast as malloc in its worst case.  When it comes to automatic testing on
Travis-CI, we can't predict the load, so I'll aim for a bound of 90%.

As before, we'll have to see if we have tests failing for reasons out of
our control.