tests/events/network_{server,client}: add
These could have been wrapped into the existing tests/events/main.c if
we used threads, but I wanted to avoid that complexity -- if we ever
need to use these files to debug problems in the event loop, it would be
easier not to worry about multithreading.

The automatic test suite only runs test 1 of the network tests, because
test 2 currently demonstrates a minor bug in event_network.c (namely
that we can't interrupt the poll()).  That fix will come in a separate
PR, which will also enable test 2.