tests: `cd` first, plus misc style fixes
Doing
	cd foo && ./cmd &
	echo $!
gives the pid of the command chain, not the pid of `cmd` itself.  This
is problematic for testing an event loop with no events which requires a
SIGUSR1 to interrupt it.

While investigating, I realized that there's no point combining those
commands -- let's start off the test function by cd'ing to the
directory.  In addition to making the upcoming empty-event-loop test
more portable, this simplifies the test scripts.