POSIX: fail if we can't work around a problem
Most of the POSIX tests have a workaround (i.e. if a header doesn't
include a symbol, we can #define it ourselves, or else omit compiling
that part of the code).

However, in the case of 'restrict', if a compiler can't handle it (even
with -std=c99), then they won't be able to compile alg/sha256.c, so
there's no point continuing with the compilation.