tests/getopt: comment reason why we didn't use usage()
The first getopt test loop uses fprintf in GETOPT_DEFAULT, while
the second one uses usage().  This should be explained,
particularly because the "canonical" example of the magic getopt
in 53d00e5bd0478fcc89bc056d192a476fb89277a4 uses usage().

Note that the second GETOPT is commented as being
"with GETOPT_MISSING_ARG this time", so that difference between the
loops is clearly intentional.