getopt: fix for gcc < 9 on solaris
Optimizers have a hard time with nonlocal gotos.  On Solaris <setjmp.h>
contains:
    #pragma unknown_control_flow(sigsetjmp)
    extern void siglongjmp(sigjmp_buf, int) __NORETURN;

The pragma is for Solaris' cc.  __NORETURN expands to
__attribute__((__noreturn__)), unless we've requested POSIX compability
(via _XOPEN_SOURCE in this case), in which case it expands to nothing.

So as far as gcc is concerned, there's a function called siglongjmp()
that has no special control flow issues.  As it happens, gcc7 and gcc8
both miscompile getopt with -O1 and have no problem with -O0 or -O2.
Later versions of gcc are fine even with -O1 and no control flow hints.

Adding __builtin_unreachable() after our siglongjmp(getopt_initloop, 1)
gives gcc enough info to produce non-broken code.