build: standardize on `test: all`
If `test:` only depends on the binary, then it can't figure out how to
build ../../cpusupport-config.h due to the "loop-back" method added in:
    2018-06-28 build: add ability to generate top-level flag files
    349592399703376673d3ccf443e501212baa2df0

Assuming that no POSIX compatibility flags or CFLAGS filtering is
needed, this is only necessary for tests which use cpusupport.  But
that's not a good assumption, and in any case it's easy to fix it
everywhere.