tests/crypto_aesctr/*.py: auto-edit main.c
Not that we'll be changing the nonce in the future or anything like that.
Rather, this is intended to reassure us that the output of
generate_test_vectors.py is actually used in main.c -- namely, run the
python script, observe that `git diff` doesn't show anything changed,
then compile & run test_aesctr knowing that the test vectors have the
same values as the python implementation.