tests/humansize: add WARNP_INIT
humansize.c includes "warnp.h", so we might expect to be able to use it
from the test.

It turns out that warnp() is only called from humansize(), which we're
not testing yet, so this isn't an actual bug.  But we might as well do
WARNP_INIT in case we ever want it for debugging, and/or adding tests
for humansize().