cpusupport: better detection of -Wno-cast-align
Versions of clang below 11 needed -Wno-cast-align to compile

    x = _mm_loadu_si128((const __m128i *)&a[0]);

but clang-11 doesn't need it.  That's a problem for the AES and SHANI
code, which does require that option.  After moving the cast into a
function, clang-11 recognizes that it needs -Wno-cast-align, which
allows it to compile the real code.

Note: in crypto_aes_aesni.c, this opearation is annotated as
    /*-
     * XXX Compiler breakage:
     * The intrinsic defined by Intel for _mm_loadu_si128 defines it as
     * taking a (const __m128i *) parameter.  This forces us to write a
     * bug: The cast to (const __m128i *) is invalid since it increases
     * the alignment requirement of the pointer.  Alas, until compilers
     * get fixed intrinsics, all we can do is code the bug and require
     * that alignment-requirement-increasing compiler warnings get
     * disabled.
     */