setuidgid: refactor string handling
This was prompted by a warning that 0-length strings weren't handled
properly.  We expect strings like "username:groupname" or "username" or
":groupname".  If somebody called
    setuidgid("", ...)
then the function would end up reading uninitialized memory.

This was not discovered in the libcperciva test suite because that
binary checked that argc == 2 and used argv[1], so we weren't passing an
empty string to that function.

Reported by:	gcc -fanalyzer