tests/crypto_aesctr: fix warmup buffer size
I made a mistake in the initial commit 6ee50068293988e6e6ab6f30cd17e648f59924fa.

The intention was to warm up with the largest buffer size (as is done in
the crc32 test).  However, the test_crc32 lists buffers going from
largest to smallest, while the test_crypto_aesctr lists buffers going
from smallest to largest.  As a result, perftest[0] is 1, not 1024.
This mistake is particularly embarrassing given that I had already added
a maxbufsize variable, precisely to avoid this type of situation.

This commit also tweaks the number of buffers.  We now process ~10 MB in
the warm-up, compared to ~34 MB for each buffer.