POSIX: check if strtod() can handle hex
Hex support for strtod() was added to the C standard in C99.
On Solaris, the documented default behaviour is not to support it [1].
(That said, `cc` and `c99` on Solaris 11.4.0.15.0 have no problem with it?)

This commit allows gcc on Solaris to recognize that it needs to use -std=c99.

[1] $ man 3 strtod
    In default mode for strtod(), only decimal, INF/INFINITY, and
    NAN/NAN(n-char-sequence) forms are recognized. In C99/SUSv3 mode,
    hexadecimal strings are also recognized.
    https://docs.oracle.com/cd/E88353_01/html/E37843/strtod-3c.html