POSIX: bail if there's runtime deviations from C99
This is a separate commit from 290d0ea901519299eac9f34c68f4fae6fd0931a3
for emphasis.

strtod() is not the only function with runtime deviation from C99 in Solaris.
The gcc [1] and clang [2] bug reports about this issue mention
deviations in logb() and freopen(), and don't mention strtod().  I'm
certain that there's other functions in the Solaris standard library
which conform to C90 instead of C99.

Our software is aimed at C99/POSIX systems.  We can make allowances for
some small violations, but there's a whole host of subtle differences
between the 30-year-old C90 and the 21-year-old C99.

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=40411
[2] https://reviews.llvm.org/D64793